# openapi/unified-openapi.yaml
openapi: 3.1.0
info:
  title: UserManagement
  version: 1.0.0
...
paths:
...
  /api/v1/meetings:schedule:
    servers:
      - url: http://localhost:33825
        description: Generated server url
    post:
      tags:
        - Meetings
      summary: Schedule a future meeting
      operationId: scheduleMeeting
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeetingManagement_ScheduleMeetingRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: >-
                  #/components/schemas/MeetingManagement_JsendResponseMeetingResponse
  /api/v1/meetings:instant:
    servers:
      - url: http://localhost:33825
        description: Generated server url
    post:
      tags:
        - Meetings
      summary: Create an instant meeting
      operationId: createInstantMeeting
      requestBody:
        content:
          application/json:
            schema:
              $ref: >-
                #/components/schemas/MeetingManagement_CreateInstantMeetingRequest
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: >-
                  #/components/schemas/MeetingManagement_JsendResponseMeetingResponse
  /api/v1/meetings/{id}:start:
    servers:
      - url: http://localhost:33825
        description: Generated server url
    post:
      tags:
        - Meetings
      summary: Start a meeting
      operationId: startMeeting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MeetingManagement_JsendResponseVoid'
...
  /api/v1/meetings/{id}:end:
    servers:
      - url: http://localhost:33825
        description: Generated server url
    post:
      tags:
        - Meetings
      summary: End a meeting
      operationId: endMeeting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MeetingManagement_JsendResponseVoid'
  /api/v1/meetings/{id}:cancel:
    servers:
      - url: http://localhost:33825
        description: Generated server url
    post:
      tags:
        - Meetings
      summary: Cancel a meeting
      operationId: cancelMeeting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MeetingManagement_JsendResponseVoid'
...
  /api/v1/meetings/{id}/settings:
    servers:
      - url: http://localhost:33825
        description: Generated server url
    patch:
      tags:
        - Meetings
      summary: Update meeting settings
      operationId: updateMeetingSettings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: >-
                #/components/schemas/MeetingManagement_UpdateMeetingSettingsRequest
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: >-
                  #/components/schemas/MeetingManagement_JsendResponseMeetingSettingsResponse
...
  /api/v1/meetings:
    servers:
      - url: http://localhost:33825
        description: Generated server url
    get:
      tags:
        - Meetings
      summary: List host meetings with cursor pagination
      operationId: listHostMeetings
      parameters:
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
        - name: pageToken
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: >-
                  #/components/schemas/MeetingManagement_JsendResponseCursorScrollResponseMeetingResponse
  /api/v1/meetings:byShortCode:
    servers:
      - url: http://localhost:33825
        description: Generated server url
    get:
      tags:
        - Meetings
      summary: Get meeting by short code
      operationId: getMeetingByShortCode
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: >-
                  #/components/schemas/MeetingManagement_JsendResponseMeetingResponse
  /api/v1/meetings/{id}:
    servers:
      - url: http://localhost:33825
        description: Generated server url
    get:
      tags:
        - Meetings
      summary: Get meeting detail by ID
      operationId: getMeeting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: >-
                  #/components/schemas/MeetingManagement_JsendResponseMeetingResponse
...
components:
  schemas:
...
    MeetingManagement_InviteeRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
      required:
        - email
    MeetingManagement_MeetingSettingsRequest:
      type: object
      properties:
        admissionPolicy:
          type: string
          minLength: 1
          pattern: ALLOW_ALL|MANUAL_APPROVAL
        joinRequestTimeoutSeconds:
          type: integer
          format: int32
          maximum: 600
          minimum: 30
        allowGuest:
          type: boolean
        muteOnEntry:
          type: boolean
        maxParticipants:
          type: integer
          format: int32
          maximum: 1000
          minimum: 1
        recordingEnabled:
          type: boolean
        screenShareMode:
          type: string
          minLength: 1
          pattern: ALL|HOST_ONLY|DISABLED
        chatEnabled:
          type: boolean
        password:
          type: string
          maxLength: 128
          minLength: 0
      required:
        - admissionPolicy
        - screenShareMode
    MeetingManagement_ScheduleMeetingRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 0
        description:
          type: string
          maxLength: 2000
          minLength: 0
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        settings:
          $ref: '#/components/schemas/MeetingManagement_MeetingSettingsRequest'
        invitees:
          type: array
          items:
            $ref: '#/components/schemas/MeetingManagement_InviteeRequest'
          maxItems: 100
          minItems: 0
      required:
        - endTime
        - settings
        - startTime
    MeetingManagement_JsendResponseMeetingResponse:
      type: object
      properties:
        status:
          type: string
        data:
          $ref: '#/components/schemas/MeetingManagement_MeetingResponse'
        message:
          type: string
    MeetingManagement_MeetingResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        hostId:
          type: string
          format: uuid
        shortCode:
          type: string
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        type:
          type: string
          enum:
            - INSTANT
            - SCHEDULED
        status:
          type: string
          enum:
            - SCHEDULED
            - LIVE
            - ENDED
            - CANCELLED
        settings:
          $ref: '#/components/schemas/MeetingManagement_MeetingSettingsResponse'
        createdAt:
          type: string
          format: date-time
    MeetingManagement_MeetingSettingsResponse:
      type: object
      properties:
        admissionPolicy:
          type: string
        joinRequestTimeoutSeconds:
          type: integer
          format: int32
        allowGuest:
          type: boolean
        muteOnEntry:
          type: boolean
        maxParticipants:
          type: integer
          format: int32
        recordingEnabled:
          type: boolean
        screenShareMode:
          type: string
        chatEnabled:
          type: boolean
        requirePassword:
          type: boolean
    MeetingManagement_CreateInstantMeetingRequest:
      type: object
      properties:
        title:
          type: string
        settings:
          $ref: '#/components/schemas/MeetingManagement_MeetingSettingsRequest'
      required:
        - settings
    MeetingManagement_JsendResponseVoid:
      type: object
      properties:
        status:
          type: string
        data: {}
        message:
          type: string
...
    MeetingManagement_JsonNullableBoolean:
      type: object
      properties:
        present:
          type: boolean
        undefined:
          type: boolean
    MeetingManagement_JsonNullableInteger:
      type: object
      properties:
        present:
          type: boolean
        undefined:
          type: boolean
    MeetingManagement_JsonNullableString:
      type: object
      properties:
        present:
          type: boolean
        undefined:
          type: boolean
    MeetingManagement_UpdateMeetingSettingsRequest:
      type: object
      properties:
        admissionPolicy:
          $ref: '#/components/schemas/MeetingManagement_JsonNullableString'
        allowGuest:
          $ref: '#/components/schemas/MeetingManagement_JsonNullableBoolean'
        muteOnEntry:
          $ref: '#/components/schemas/MeetingManagement_JsonNullableBoolean'
        maxParticipants:
          $ref: '#/components/schemas/MeetingManagement_JsonNullableInteger'
        recordingEnabled:
          $ref: '#/components/schemas/MeetingManagement_JsonNullableBoolean'
        screenShareMode:
          $ref: '#/components/schemas/MeetingManagement_JsonNullableString'
        chatEnabled:
          $ref: '#/components/schemas/MeetingManagement_JsonNullableBoolean'
        joinRequestTimeoutSeconds:
          $ref: '#/components/schemas/MeetingManagement_JsonNullableInteger'
    MeetingManagement_JsendResponseMeetingSettingsResponse:
      type: object
      properties:
        status:
          type: string
        data:
          $ref: '#/components/schemas/MeetingManagement_MeetingSettingsResponse'
        message:
          type: string
    MeetingManagement_CursorScrollResponseMeetingResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/MeetingManagement_MeetingResponse'
        size:
          type: integer
          format: int32
        nextPageToken:
          type: string
    MeetingManagement_JsendResponseCursorScrollResponseMeetingResponse:
      type: object
      properties:
        status:
          type: string
        data:
          $ref: >-
            #/components/schemas/MeetingManagement_CursorScrollResponseMeetingResponse
        message:
          type: string
...